Convert parser to TypeScript
